+++ solnlib/modular_input/event_writer.py
@@ -83,7 +83,7 @@ class EventWriter(metaclass=ABCMeta):
            >>>     host='localhost',
            >>>     source='Splunk',
            >>>     sourcetype='misc',
-           >>>     fields='{'accountid': '603514901691', 'Cloud': u'AWS'}'
+           >>>     fields={'accountid': '603514901691', 'Cloud': u'AWS'},
            >>>     stanza='test_scheme://test',
            >>>     unbroken=True,
            >>>     done=True)
